Project restructure into packages
For some reason tests have problems. Need fixing.
