+++ dnp/__init__.py
@@ -1,0 +1,3 @@
+from .entity import Enemy, Hero, Weapon, Spell
+from .logic import Dungeon, Fight
+from .utils import utils
