+++ dnp/entity/Spell.py
