+++ dnp/entity/Weapon.py
