+++ dnp/entity/__init__.py
@@ -1,0 +1,4 @@
+from .Enemy import Enemy
+from .Hero import Hero
+from .Weapon import Weapon
+from .Spell import Spell
