+++ dnp/logic/__init__.py
@@ -1,0 +1,2 @@
+from .Dungeon import Dungeon
+from .Fight import Fight
