+++ maps/basic_loot_list_example.json
