+++ media/ascii_griffin.txt
