+++ test/__init__.py
@@ -1,0 +1,10 @@
+from .test_Enemy import test_Enemy
+from .test_Hero import Test_hero
+from .test_Weapon import Test_weapon
+from .test_Spell import test_Spell
+
+
+from .test_Dungeon import test_Dungeon
+from .test_Fight import TestFight
+
+from .test_utils import test_utils
